di: implement di::Function, a move-only type erased function container
This implementation is essentially std::move_only_function, with the
addition of nontype callable support, as well as additional factory
functions which are required to be called when creating large function
objects.

Any type which is less than 2 void*'s in size, will be stored inline
inside the function object (assuming it is move constructible). This
means that function pointers, reference wrappers, and lambdas which
capture only the this pointer will all be stored without requiring
allocations. Additionally, the nontype support allows creating functions
directly from member pointers, member functions, and static function pointers
without even needing to store the nontype directly.

The Function type is only constructible if doing so is noexcept, so, the
stored value type must be stored inline. This is fine since almost every
relevant case which requires storing a type erased function with either
use a function pointer, lambda capturing this, or a nontype constructor.
Lambdas which capture a large amount of state will usually not need to be
stored, and so can be passed using FunctionRef.

If a large function is actually needed, the make_function() and
try_make_function() factories allow for allocation and the try_*
variant additionally returns a di::Result<> instead of panicking on
failure.