di: implement a to_owned() customization point
This allows code to easily convert from a view type, such as StringView
or PathView or Span, into an owning type, like String, Path, Vector, or
Array. This solution is a lot more ergonomic than calling di::create<>()
with the proper type, since the borrowed type "knows" its owned conterpart,
which means explicit types do not need to be listed.