dius: handle assertion failures properly in Iros
Instead of catching signals to determine that as assertion failed, the
test runner will simply hook into the assertion handler and get notified
when failures occurred. When running on Linux, if a segmentation fault
happens, the test runner will report the failure and exit without
running the test of the tests. This is because the heap could easily be
corrupted, and we have no way to fix it.