iris: pass arguments to tasks
The program's arguments and environment can now be set before task
creation, and the task will receive them when it starts. This
information is recorded in an immutable reference counted data
datastructure, and so can be shared between tasks (threads) and also
used for some basic task introspection in the future.