dius: test process argument passing
This ensures that the dius::process::Process abstraction works on both
Linux and Iros. As a side effect, this tests that kernel's new argument
passing works as intended.

This patch also adds a posix_spawn backend to dius::process::Process,
which is used when not using the custom runtime. This is necessary to
run sanitizers and coverage tools properly.