di: get rid of circular dependency in ASSERT macro
Essentially, in order for the ASSERT macro to pretty-print the
expression, it needs to stringify the arguments. This requires the
entire formatting machinery to be available, which in turn requires the
string class, which in turn requires memory allocation, and so on. It
turns out these components also want to use the ASSERT macro.

The old solution was to have a separate header that was included by the
public prelude.h header, but this requires essentially making the entire
library into a single header. This results in terrible compile times.

The correct solution is to get rid of this circular dependency, so that
headers can be included like normal. However, this makes stringify the
arguments impossible. This would be doable if we could declare a
templated function out-of-line, but this isn't possible in C++ (maybe
with modules?).

Instead, we can define an out-of-line function which takes the
stringified arguments preconverted to a null-terminated string array.
This has 0 c++ dependencies. This allows the kernel and dius to have
their own custom assert handlers, which can still pretty-print the
arguments in general.

However, within di, these pretty-printed assertion macros cannot be
used. Instead, the library must use the plain DI_ASSERT() macro, which
cannot pretty-print the arguments, but can at least capture the source
location. This is a reasonable compromise, as any alternative would rely
on the same circular dependency, most likely with some insane
preprocessor hacks.

This enables future commits to remove the global prelude.h header, and
force clients to include what they use.