unicode: implement WIP grapheme clusterer and working view using it
This is a proof of concept for a fully correct implementation of
grapheme clustering. The key component is a DFA with state transitions
mapped to a lookup table at compile time. We use this in C++ 20
range style view (based on std::split) to actually segment graphemes.
For usage in ttx, we will use the GraphemeClusterer DFA directly.